Fixed Build Problem
-Fixed problem of having two versions of the same function in the object class.
-Fixed a double to float conversion warning.

+++ SimpleActionRpg/Vector2D.cpp
@@ -148,5 +148,5 @@ float Vector2D::GetLength()
 //*********************************************************************************************************************************************
 void Vector2D::CalculateLength()
 {
-   mLength = sqrt(powf(fabsf(mComponentX), 2.0F) + powf(fabsf(mComponentY), 2.0F));
+   mLength = sqrtf(powf(fabsf(mComponentX), 2.0F) + powf(fabsf(mComponentY), 2.0F));
 }
